Issues Resolved

+++ src/main/java/com/stackroute/movie/service/MovieServiceImpl.java
@@ -11,13 +11,13 @@ import java.util.List;
 
 @Service
 public class MovieServiceImpl implements MovieService {
-    MovieRepository movieRepository;
+    private MovieRepository movieRepository;
     @Autowired
     public MovieServiceImpl(MovieRepository movieRepository) {
         this.movieRepository = movieRepository;
     }
     @Override
-    public Movie saveMovie(Movie movie) throws MovieAlreadyExistsException {
+    public Movie saveMovie(Movie movie) throws MovieAlreadyExistsException, Exception {
         if(movieRepository.existsById(movie.getMovieId()))
         {
             throw new MovieAlreadyExistsException("Movie already exists");
@@ -30,21 +30,37 @@ public class MovieServiceImpl implements MovieService {
         return savedMovie;
     }
     @Override
-    public List<Movie> getAllMovies() {
+    public List<Movie> getAllMovies() throws Exception {
         return movieRepository.findAll();
     }
     @Override
-    public void deleteMovie(int movieId) {
-        movieRepository.deleteById(movieId);
+    public Movie deleteMovie(int movieId) throws MovieNotFoundException, Exception {
+        if(movieRepository.existsById(movieId))
+        {
+            Movie tmovie = movieRepository.getOne(movieId);
+            movieRepository.deleteById(movieId);
+            return tmovie;
+        }
+        else
+        {
+            throw new MovieNotFoundException("Movie not Found");
+        }
     }
     @Override
-    public void updateMovie(Movie movie) {
-        movieRepository.deleteById(movie.getMovieId());
-        movieRepository.save(movie);
+    public void updateMovie(Movie movie) throws MovieNotFoundException, Exception {
+        if(movieRepository.existsById(movie.getMovieId()))
+        {
+            movieRepository.deleteById(movie.getMovieId());
+            movieRepository.save(movie);
+        }
+        else
+        {
+            throw new MovieNotFoundException("Movie Not Found");
+        }
     }
 
     @Override
-    public List<Movie> findByName(String name) throws MovieNotFoundException {
+    public List<Movie> findByName(String name) throws MovieNotFoundException, Exception {
         if(movieRepository.findByName(name).isEmpty())
         {
             throw new MovieNotFoundException("Movie Not Found");
